[tokens-page] Moved TokenModal to Tokens folder and also added Login link to Sidebar
