+++ client/src/common/components/Sidebar/Sidebar.tsx
@@ -8,6 +8,7 @@ import {
     faDownload,
     faFileExport,
     faHome,
+    faRightToBracket,
 } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useRouter } from "next/router";
@@ -150,6 +151,19 @@ export const Sidebar = (): JSX.Element => {
                 >
                     <FontAwesomeIcon className="my-auto pe-1" icon={faCoins} />
                     <span className="fw-bold">{"Tokens"}</span>
+                </span>
+                <span
+                    className={`${styles.sidebar_individual_content} d-flex flex-row justify-content-center pt-2 border-bottom pb-2`}
+                    onClick={async (): Promise<void> => {
+                        await router.push("/login");
+                    }}
+                    role="button"
+                >
+                    <FontAwesomeIcon
+                        className="my-auto pe-1"
+                        icon={faRightToBracket}
+                    />
+                    <span className="fw-bold">{"Login"}</span>
                 </span>
             </div>
         </>
